requirements.txt
streamlit>=1.28.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0

